🚚 Move Position into its own module
I'm trying to make `input::mod` smaller and keep each entity in their own module. It's much more maintainable.
