+++ src/tests/separated_by1.rs
@@ -1,5 +1,5 @@
-use crate::input::Position;
 use crate::*;
+use input::position::Position;
 
 #[test]
 fn empty() {
